messages don't activate
I didn't realize that "activateFunc" and "performFunc" are different things
in the C implementation. The Activate method of this implementation's
objects corresponds to the latter, not the former. Messages have only the
former, but I still gave them a non-trivial Activate method, which was
incorrect and caused expressions like Message fromString to fail.