implemented Object activation
The original implementation's means of having objects activate was to
use the activateFunc field of the object's tag, essentially a virtual
method that could be changed at run-time. Activation was then a matter
of executing the object's activateFunc if it was non-NULL, and it meant
that objects did not need a slot to denote activatability. In this
implementation, I've decided instead to have isActivatable be a slot
with a value of true or false. It seems very inefficient to have an
Activatable attribute of every object, especially since the Object type
is embedded in every primitive type, which can't be set activatable.
Instead, we have a slot that is only on the objects which need it, at
the cost of a slot lookup per object activation, which has very little
impact.