speed improvements
Moved GetSlot to be a VM method to allow it to use a reusable set object instead of map[Interface]struct{}, and also so that contextWithSlot returns the correct type of object when the receiver has the slot.

Removed the Raw interface, because asserting to Raw in GetSlot was slow. Its methods are now in Interface.

The primary bottlenecks still present in GetSlot are actually looking up the slot in the slots map and acquiring and releasing the object lock, each taking about 20% of the execution time in my profiles. I don't plan to change any of these.

Also removed the VM.NumberMemo map in favor of other memoization structures. A new mknumbermemo program generates methods to initialize them.