implement new addon loading system
This is the first half of the second half of the final API rework. I developed this on my Windows computer, so I can't test it at the moment. Right now, this push is to let me try loading Range on my Chromebook.

Addons in this system currently must be built as Go plugins, even if they contain only Io code. cmd/mkaddon is a program that generates the addon glue code from a YAML manifest and initialization scripts in Io. addons/Range provides an example of how to set it up, including a highly commented manifest.

Pure Io addons won't be supported in the future. If they were available, then I'd still ignore them when plugins are not supported or disabled to keep consistent behavior. That means that pure Io addons would have to run mkaddon to generate a package anyway for static builds.

I plan to make another command that will scan Go files for functions assignable to iolang.Fn to auto-generate function definitions in the manifest. Currently, that is by far the most tedious and laborious part of making a new addon, and it is fairly simple to automate.

The primary problem at this point is that mkaddon requires that the package name for an addon be CamelCase, so lint complains. However, that requirement is actually beneficial for the only currently existing addon, because it can't be named "range".