safer UniqueID implementation
Using object pointers as unique IDs is unsafe because the garbage collector can move objects during function calls. In pathological cases, it would be possible to construct an infinite loop in GetSlot by exploiting this. Using a global counter for a UniqueID field on each object is safer, and it means we don't need a separate, slow implementation for situations in which we can't import unsafe.