make testutils subpackage
The testing utils are very useful, but it isn't ideal to include them in every build of iolang. Fully moving them to a separate package would cause circular imports, so we'd have to break out absolutely everything into separate packages, which would be a serious hit to usability as well as a massive effort. Instead, just copy everything.