make System exit cause a graceful shutdown
There is a new Stop value, ExitStop, which nothing should handle. (This could be used to provide atexit-like behavior, though.) System exit causes the scheduler to send this Stop to every coroutine, so they will stop after finishing their current message.

This System exit implementation probably ought to be opt-in for sandboxes, not the one used for everything, because it will wait for blocking coroutines before exiting.