optimize for objects with only one proto
The overwhelming majority of objects have only one proto. To specialize for this case, each object's first proto is now its own field, and the remainder go in a slice, which is nil if the object has only one proto. This also means GetSlot can avoid using the proto stack for objects that have only one proto, saving some slice operations.

Most benchmarks see improvements in the range of 2.5% to 5%. EvalParallel instead is about 3% slower.