separate edges from nodes on slots tries
Trie nodes are now of type slotBranch instead of slotRecord; the latter now only includes the trie's nonzero edges.

The principal motivation for this change is to allow slot names containing NUL characters to have the same synchronization semantics as all other slots. The previous implementation, where the first child from a node was always the zero edge, made it impossible to know when to spin for such edges. This change adds a separate atomic flag at the branch level for the zero edge.

A side-effect is that branches with many edges are now smaller, as only the branch itself includes space for a leaf. Additionally, benchmarks appear to be very slightly faster.